Core.Accounts.User (core v0.1.0)

The User module encapsulates the changesets for creating and updating users.

The logic for hashing passwords is also managed here. Bcrypt is used as adviced by OWASP:

https://cheatsheetseries.owasp.org/cheatsheets/Password_Storage_Cheat_Sheet.html#password-hashing-algorithms

Functions

confirm_changeset(user)

Confirms the account by setting confirmed_at.

display_changeset(user, attrs)

A user changeset for changing the display name.

email_changeset(user, attrs)

A user changeset for changing the email.

It requires the email to change otherwise an error is added.

get_gender(map)

label(map)

new_session_changeset(credentials)

password_changeset(user, attrs, opts \\ [])

A user changeset for changing the password.

Options

:hash_password - Hashes the password so it can be stored securely in the database and ensures the password field is cleared to prevent leaks in the logs. If password hashing is not needed and clearing the password field is not desired (like when using this changeset for validations on a LiveView form), this option can be set to false. Defaults to true.

registration_changeset(user, attrs, opts \\ [])

A user changeset for registration.

It is important to validate the length of both email and password. Otherwise databases may truncate the email without warnings, which could lead to unpredictable or insecure behaviour. Long passwords may also be very expensive to hash for certain algorithms.

Options

:hash_password - Hashes the password so it can be stored securely in the database and ensures the password field is cleared to prevent leaks in the logs. If password hashing is not needed and clearing the password field is not desired (like when using this changeset for validations on a LiveView form), this option can be set to false. Defaults to true.

sso_changeset(user, attrs)

A user changeset for use with Single Sign On systems. It implicitly trusts most data that it receives.

user_profile_changeset(user, attrs)

User changeset for profile page

valid_email_changeset(email \\ nil)

valid_password?(arg1, password)

Verifies the password.

If there is no user or the user doesn't have a password, we call Bcrypt.no_user_verify/0 to avoid timing attacks.

validate_current_password(changeset, password)

Validates the current password otherwise adds an error to the changeset.

visited_changeset(user, attrs)