Core.Sentry.Gen (core v0.1.0)

Sends a test event to Sentry to check if your Sentry configuration is correct.

This task reports a RuntimeError exception like this one:

%RuntimeError{message: "Testing sending Sentry event"}

Options

compile - boolean, use to does not compile, even if files require compilation.

type - exception or message. Defaults to exception. Available since v10.1.0..

stacktrace - does not include a stacktrace in the reported event. Available since v10.1.0..

Functions

test_event()