Core.Studies (core v0.1.0)

The Studies context.

Functions

add_author(study, researcher)

add_owner!(study, user)

assign_owners(study, users)

change_study(study, attrs \\ %{})

Returns an %Ecto.Changeset{} for tracking study changes.

Examples

iex> change_study(study)
%Ecto.Changeset{data: %Study{}}

create_study(changeset, researcher)

Creates a study.

delete_study(study)

Deletes a study.

Examples

iex> delete_study(study)
{:ok, %Study{}}

iex> delete_study(study)
{:error, %Ecto.Changeset{}}

get_study!(id)

Gets a single study.

Raises Ecto.NoResultsError if the Study does not exist.

Examples

iex> get_study!(123)
%Study{}

iex> get_study!(456)
** (Ecto.NoResultsError)

get_study_changeset(attrs \\ %{})

list_authors(study)

list_owned_studies(user, opts \\ [])

Returns the list of studies that are owned by the user.

list_owners(study)

list_studies(opts \\ [])

Returns the list of studies.

Examples

iex> list_studies()
[%Study{}, ...]

list_studies_with_published_survey(opts \\ [])

list_subject_studies(user, opts \\ [])

Returns the list of studies where the user is a subject.

list_survey_tools(study)

update_study(changeset)

Updates a study.

Examples

iex> update_study(study, %{field: new_value})
{:ok, %Study{}}

iex> update_study(study, %{field: bad_value})
{:error, %Ecto.Changeset{}}

update_study(study, attrs)