Core.Survey.Tools (core v0.1.0)

Survey tools allow a researcher to setup a link to an external survey tool. The participant goes through the flow described below:

Receive invitation to start a survey (mail, push etc.).

Open survey tool, this opens it on the platform and requires authentication.

The participant is then redirected to the survey at a 3rd party web-application.

After completion the user is redirect back to the platform.

The platform registers the completion of this survey for the participant.

A researcher is required to configure the 3rd party application with a redirect link. The redirect link to be used is show on the survey tool configuration screen (with copy button).

IDEA: The tool requires a sucessful round-trip with a verify flow to ensure that everything is configured correctly.

Participants need to be invited to a particular survey explicitly. This avoids the situation where a new user joins a study and then can immediately complete previous surveys.

Once a participant has completed a survey they are no longer allowed to enter it a second time. The status is clearly shown when the attempt to do so.

IDEA: A list of surveys can be access by the notification icon which is shown on all screens.

Functions

apply_participant(survey_tool, user)

change_survey_tool(survey_tool, type, attrs \\ %{})

Returns an %Ecto.Changeset{} for tracking survey_tool changes.

complete_task!(task)

count_completed_tasks(survey_tool)

count_pending_tasks(survey_tool)

count_tasks(survey_tool, status_list)

create_survey_tool(attrs, study, promotion, content_node)

Creates a survey_tool.

create_task(survey_tool, user)

delete_survey_tool(survey_tool)

Deletes a survey_tool.

delete_task(task)

get_by_promotion(promotion_id)

get_or_create_task(survey_tool, user)

get_or_create_task!(survey_tool, user)

get_survey_tool(id)

get_survey_tool!(id)

Gets a single survey_tool.

Raises Ecto.NoResultsError if the Survey tool does not exist.

get_task(survey_tool, user)

list_participants(survey_tool)

list_participants_without_task(survey_tool)

list_participations(user)

list_survey_tools()

Returns the list of survey_tools.

list_tasks(survey_tool)

participant?(survey_tool, user)

Specs

participant?(
  atom() | %{:id => any(), optional(any()) => any()},
  atom() | %{:id => any(), optional(any()) => any()}
) :: boolean()

setup_tasks_for_participants!(participants, survey_tool)

update_survey_tool(changeset)

update_survey_tool(survey_tool, type, attrs)

Updates a survey_tool.

withdraw_participant(survey_tool, user)