CoreWeb.Cldr.Date (core v0.1.0)

Functions

to_string(date, options \\ [])

Specs

to_string(map(), Keyword.t()) ::
  {:ok, String.t()} | {:error, {module(), String.t()}}

Formats a date according to a format string as defined in CLDR and described in TR35

Arguments

date is a %Date{} struct or any map that contains the keys year, month, day and calendar

options is a keyword list of options for formatting. The valid options are:

Options

format: :short | :medium | :long | :full or a format string. The default is :medium

locale: any locale returned by Cldr.known_locale_names/1. The default is Cldr.get_locale().

number_system: a number system into which the formatted date digits should be transliterated

Returns

{:ok, formatted_string} or

{:error, reason}

Examples

iex> CoreWeb.Cldr.Date.to_string ~D[2017-07-10], format: :medium, locale: "en"
{:ok, "Jul 10, 2017"}

iex> CoreWeb.Cldr.Date.to_string ~D[2017-07-10], locale: "en"
{:ok, "Jul 10, 2017"}

iex> CoreWeb.Cldr.Date.to_string ~D[2017-07-10], format: :full, locale: "en"
{:ok, "Monday, July 10, 2017"}

iex> CoreWeb.Cldr.Date.to_string ~D[2017-07-10], format: :short, locale: "en"
{:ok, "7/10/17"}

iex> CoreWeb.Cldr.Date.to_string ~D[2017-07-10], format: :short, locale: "fr"
{:ok, "10/07/2017"}

iex> CoreWeb.Cldr.Date.to_string ~D[2017-07-10], format: :long, locale: "af"
{:ok, "10 Julie 2017"}

to_string!(date, options \\ [])

Specs

to_string!(map(), Keyword.t()) :: String.t() | no_return()

Formats a date according to a format string as defined in CLDR and described in TR35

Arguments

date is a %Date{} struct or any map that contains the keys year, month, day and calendar

options is a keyword list of options for formatting.

Options

format: :short | :medium | :long | :full or a format string. The default is :medium

locale is any valid locale name returned by Cldr.known_locale_names/0 or a Cldr.LanguageTag struct. The default is Cldr.get_locale/0

number_system: a number system into which the formatted date digits should be transliterated

Returns

formatted_date or

raises an exception.

Examples

iex> CoreWeb.Cldr.Date.to_string! ~D[2017-07-10], format: :medium, locale: "en"
"Jul 10, 2017"

iex> CoreWeb.Cldr.Date.to_string! ~D[2017-07-10], locale: "en"
"Jul 10, 2017"

iex> CoreWeb.Cldr.Date.to_string! ~D[2017-07-10], format: :full,locale: "en"
"Monday, July 10, 2017"

iex> CoreWeb.Cldr.Date.to_string! ~D[2017-07-10], format: :short, locale: "en"
"7/10/17"

iex> CoreWeb.Cldr.Date.to_string! ~D[2017-07-10], format: :short, locale: "fr"
"10/07/2017"

iex> CoreWeb.Cldr.Date.to_string! ~D[2017-07-10], format: :long, locale: "af"
"10 Julie 2017"