CoreWeb.Cldr.Time (core v0.1.0)

Functions

to_string(time, options \\ [])

@spec to_string(map(), Keyword.t()) ::
  {:ok, String.t()} | {:error, {module(), String.t()}}

Formats a time according to a format string as defined in CLDR and described in TR35

Returns

{:ok, formatted_time} or

{:error, reason}.

Arguments

time is a %DateTime{} or %NaiveDateTime{} struct or any map that contains the keys hour, minute, second and optionally calendar and microsecond

options is a keyword list of options for formatting.

Options

format: :short | :medium | :long | :full or a format string. The default is :medium

locale: any locale returned by Cldr.known_locale_names/1. The default is Cldr.get_locale()

number_system: a number system into which the formatted date digits should be transliterated

era: :variant will use a variant for the era is one is available in the locale. In the "en" locale, for example, era: :variant will return "BCE" instead of "BC".

period: :variant will use a variant for the time period and flexible time period if one is available in the locale. For example, in the "en" locale period: :variant will return "pm" instead of "PM"

Examples

iex> Cldr.Time.to_string ~T[07:35:13.215217]
{:ok, "7:35:13 AM"}

iex> Cldr.Time.to_string ~T[07:35:13.215217], format: :short
{:ok, "7:35 AM"}

iex> Cldr.Time.to_string ~T[07:35:13.215217], format: :medium, locale: "fr"
{:ok, "07:35:13"}

iex> Cldr.Time.to_string ~T[07:35:13.215217], format: :medium
{:ok, "7:35:13 AM"}

iex> {:ok, datetime} = DateTime.from_naive(~N[2000-01-01 23:59:59.0], "Etc/UTC")
iex> Cldr.Time.to_string datetime, format: :long
{:ok, "11:59:59 PM UTC"}

to_string!(time, options \\ [])

@spec to_string!(map(), Keyword.t()) :: String.t() | no_return()

Formats a time according to a format string as defined in CLDR and described in TR35.

Arguments

time is a %DateTime{} or %NaiveDateTime{} struct or any map that contains the keys hour, minute, second and optionally calendar and microsecond

options is a keyword list of options for formatting.

Options

format: :short | :medium | :long | :full or a format string. The default is :medium

locale is any valid locale name returned by Cldr.known_locale_names/0 or a Cldr.LanguageTag struct. The default is Cldr.get_locale/0

number_system: a number system into which the formatted date digits should be transliterated

era: :variant will use a variant for the era is one is available in the locale. In the "en" locale, for example, era: :variant will return "BCE" instead of "BC".

period: :variant will use a variant for the time period and flexible time period if one is available in the locale. For example, in the "en" locale period: :variant will return "pm" instead of "PM"

Returns

formatted_time_string or

raises an exception.

Examples

iex> CoreWeb.Cldr.Time.to_string! ~T[07:35:13.215217]
"7:35:13 AM"

iex> CoreWeb.Cldr.Time.to_string! ~T[07:35:13.215217], format: :short
"7:35 AM"

iex> CoreWeb.Cldr.Time.to_string ~T[07:35:13.215217], format: :short, period: :variant
{:ok, "7:35 AM"}

iex> CoreWeb.Cldr.Time.to_string! ~T[07:35:13.215217], format: :medium, locale: "fr"
"07:35:13"

iex> CoreWeb.Cldr.Time.to_string! ~T[07:35:13.215217], format: :medium
"7:35:13 AM"

iex> {:ok, datetime} = DateTime.from_naive(~N[2000-01-01 23:59:59.0], "Etc/UTC")
iex> CoreWeb.Cldr.Time.to_string! datetime, format: :long
"11:59:59 PM UTC"