Frameworks.GreenLight.PermissionMap (core v0.1.0)

The permission map holds a mapping of permissions to roles. The permission map is intended to be an implementation detail of the authorization framework. Convenience functions to build it are provided in other modules.

Types

permission()

@type permission() :: binary()

role()

@type role() :: atom()

role_enum()

@type role_enum() :: role_set() | role_list()

role_list()

@type role_list() :: [role()]

role_set()

@type role_set() :: MapSet.t(role())

t()

@opaque t()

Functions

allowed?(permission_map, permission, principal_roles)

@spec allowed?(t(), permission(), role() | role_list()) :: boolean()

@spec allowed?(t(), binary(), role_set()) :: boolean()

Returns wheter or not a the given set of roles should be allowed for the given permission.

grant(permission_map, permission, role)

@spec grant(t(), permission(), role() | role_list()) :: t()

Add a mapping for the given role and permission if does not already exist.

list_permission_assignments(permission_map)

merge(a, b)

@spec merge(t(), t()) :: t()

Merge two permission maps. The result of the merge is a permission map for which a permission that is available in either (or both) maps has the combined roles of both permission maps.

new()

@spec new() :: t()

new(mapping)

@spec new(%{required(permission()) => role_enum()}) :: t()

new_t(permission_map)

@spec new_t(any()) :: t()

This constructor is here to appease Dialyzer when a PermissionMap is used as a module attribute.

roles(permission_map, permission)

@spec roles(t(), permission()) :: role_set()

Return a MapSet of roles for a given permission. This always returns a MapSet (might be empty).