Systems.Alliance.ToolModel (core v0.1.0)

The alliance tool schema.

Functions

changeset(tool, params)

changeset(tool, arg2, params)

external_path(arg1, next_id)

operational_changeset(entity, attrs)

preload_graph(fields)

Callback implementation for Frameworks.Utility.Schema.preload_graph/1.

prepare_url(url, replacements)

ready?(tool)

validate(changeset)

validate(changeset, atom)

validate_query_string(query, field)

validate_url(changeset, field, options \\ [])

variable_parser(binary, opts \\ [])

@spec variable_parser(binary(), keyword()) ::
  {:ok, [term()], rest, context, line, byte_offset}
  | {:error, reason, rest, context, line, byte_offset}
when line: {pos_integer(), byte_offset},
     byte_offset: pos_integer(),
     rest: binary(),
     reason: String.t(),
     context: map()

Parses the given binary as variable_parser.

Returns {:ok, [token], rest, context, position, byte_offset} or {:error, reason, rest, context, line, byte_offset} where position describes the location of the variable_parser (start position) as {line, offset_to_start_of_line}.

To column where the error occurred can be inferred from byte_offset - offset_to_start_of_line.

Options

:byte_offset - the byte offset for the whole binary, defaults to 0

:line - the line and the byte offset into that line, defaults to {1, byte_offset}

:context - the initial context value. It will be converted to a map