Systems.Assignment.Public (core v0.1.0)

The assignment context.

Functions

add_participant!(model, user)

apply_member(id, user, identifier, reward_amount)

assign_tester_role(tool, user)

attention_list_enabled?(map)

cancel(assignment, user)

complete_task(crew, identifier)

copy(assignment, info, workflow, budget, auth_node)

copy_info(info)

copy_workflow(workflow)

count_participants(assignment)

count_participants_finished(assignment)

create_page_ref(assignment, key)

decline_member(model, user)

delete_consent_agreement(assignment)

delete_page_ref(page_ref)

delete_storage_endpoint!(assignment)

exclude(assignment1, assignment2)

excluded?(assignment, map)

Is user excluded? from joining given assignment

expiration_timestamp(map)

expired_user_assignments(from)

get(id, preload \\ [])

get!(id, preload \\ [])

get_by(association, preload \\ [])

get_by(field_name, id, preload)

get_by_content_page(page, preload \\ [])

get_by_tool(tool, preload \\ [])

get_by_tool_ref(workflow, preload \\ [])

get_by_workflow_item_id(workflow_item_id, preload \\ [])

get_crew(assignment)

get_task(tool, identifier)

get_workflow!(id, preload \\ [])

has_open_spots?(assignment)

Is assignment open for new members?

idempotence_key(assignment_id, user_id)

include(assignment1, assignment2)

list_by_crew(crew, preload \\ [])

list_by_participant(user, preload \\ [])

list_participants(assignment)

Lists the participants of the assignment. Returns a list of maps with the following keys:

user_id

public_id

external_id

member_id

list_signatures(assignment)

list_tasks(model)

list_user_ids(assignment_ids)

mark_expired_debug(assignment, force)

member?(map, user)

open_spot_count(map)

How many new members can be added to the assignment?

owner!(assignment)

participant_id(model, user)

payout_participant(model, user)

prepare(attrs, crew, info, page_refs, workflow, budget, consent_agreement, auth_node)

prepare_info(attrs)

prepare_page_ref(auth_node, key)

prepare_page_refs(template, auth_node)

prepare_tool_ref(special, tool)

prepare_workflow(special, items, type, auth_node)

prepare_workflow_item(tool_ref, attrs \\ %{})

prepare_workflow_items(tool_refs)

query_by_tool(field_name, id, preload)

query_by_tool_ref(tool_ref_id, preload)

ready?(map)

reject_task(assignment, task, rejection)

reset_member(assignment, user, opts \\ [])

rewarded_amount(model, user)

rollback_deposit(multi, assignment, user)

rollback_expired_deposits()

rollback_expired_deposits(from)

run_apply_member(crew, user, identifier, expire_at)

search_subject(tool, user)

start_task(tool, identifier)

status(map, user)

task_labels(map)

tester?(arg1, user_ref)

timestamp(arg1)

timestamp(map, user)

update(assignment, attrs)

update!(assignment, attrs)

update_budget(assignment, budget)

update_consent_agreement(assignment, consent_agreement)

update_storage_endpoint(assignment, storage_endpoint)