Systems.Campaign.Public (core v0.1.0)

The Studies context.

Functions

add_author(campaign, researcher)

add_owner!(campaign, user)

all(ids, preload \\ [])

assign_coordinators(campaign)

assign_owners(campaign, users)

change(campaign, attrs \\ %{})

completed?(campaign)

copy(authors, campaign)

copy(campaign, promotion, assignment, submissions, auth_node)

create(promotion, assignment, submissions, researcher, auth_node)

Creates a campaign.

delete(id)

get!(id, preload \\ [])

get_by_promotable(promotable, preload \\ [])

get_by_promotion(promotion, preload \\ [])

get_by_submission(submission, preload \\ [])

get_changeset(attrs \\ %{})

get_user_pools_by_currency(user, currency)

handle_exclusion(assignment, items)

import_student_reward(student_id, amount, session_key, currency)

import_student_reward_exists?(student_id, session_key)

list(opts \\ [])

list_authors(map)

list_by_budget(model, preload \\ [])

list_by_pools_and_submission_status(pools, submission_status, opts \\ [])

list_by_promotables(promotable_ids, preload)

list_excluded_campaigns(campaigns, preload \\ [])

list_excluded_user_ids(campaign_ids)

list_owned_campaigns(user, opts \\ [])

Returns the list of studies that are owned by the user.

list_owners(campaign, preload \\ [])

list_subject_campaigns(user, opts \\ [])

Returns the list of studies where the user is a subject.

list_submitted(pool, opts \\ [])

Returns the list of campaigns submitted at least once.

list_tools(campaign)

list_tools(campaign, schema)

mark_expired_debug(force \\ false)

Marks expired tasks in online campaigns based on updated_at and estimated duration. If force is true (for debug purposes only), all pending tasks will be marked as expired.

mark_expired_debug(map, force)

Marks expired tasks in given campaign

open_spot_count(arg1)

original_author(campaign)

payout_participant(assignment, user)

query_user_pools_by_currency(user, currency)

ready?(id)

remove_owner!(campaign, user)

reward_amount(model)

reward_value(assignment)

rewarded_amount(assignment, user)

submission_updated(arg1)

sync_student_credits()

Synchronizes accepted student tasks with bookkeeping.

update(changeset)

Updates a get_campaign.

Examples

iex> update(get_campaign, %{field: new_value})
{:ok, %Campaign.Model{}}

iex> update(get_campaign, %{field: bad_value})
{:error, %Ecto.Changeset{}}

update(campaign, attrs)

update_coordinator_role(user, value)

user_has_currency?(user, currency)

validate_open(assignment, user)