Systems.DataDonation.Public (core v0.1.0)

A data donation allows a researcher to ask participants to submit data. This data is submitted in the form of a file that is stored on the participants device.

Tools are provided that allow for execution of filtering code on the device of the participant. This ensures that only the data that is needed for the study is shared with the researcher.

Functions

add_task(tool, task_type)

copy(tool, auth_node)

create(attrs, auth_node)

create_task(tool, position, task_type, special)

create_task_special(atom)

delete(tool)

get_document_task!(id, preload \\ [])

get_task(tool_id, position, preload \\ [])

get_task!(id, preload \\ [])

get_tool!(id, preload \\ [])

list()

list_tasks(tool_id, preload \\ [])

switch_position(task1, task2)

task_count(tool_model)

update(changeset)