Systems.Email.Factory (core v0.1.0)

Functions

account_confirmation_instructions(user, url)

account_created(user)

already_activated_notification(user, url)

debug(subject, message, from_user, to_user)

mail_user(emails)

notification(title, byline, message, to)

render(email, template, assigns \\ [])

Render an Phoenix template and set the body on the email.

Pass an atom as the template name (:welcome_email) to render HTML and plain text emails. Use a string if you only want to render one type, e.g. "welcome_email_text" or "welcome_email_html". Scroll to the top for more examples.

reset_password_instructions(user, url)

update_email_instructions(user, url)