Systems.Graphite.Gen (core v0.1.0)

Functions

create_submissions(leaderboard_id, amount, prefix)

Create amount submissions for the specified leaderboard, where the submission description is created with the prefix, followed by a number.

delete_submissions(prefix)

Delete the submissions where the submission description starts with prefix.