Systems.Advert.Public (core v0.1.0)

The Studies context.

Functions

all(ids, preload \\ [])

assign_owners(advert, users)

change(advert, attrs \\ %{})

completed?(advert)

copy(advert, promotion, assignment, submission, auth_node)

create(promotion, assignment, submission, researcher, auth_node)

Creates a advert.

delete(id)

get!(id, preload \\ [])

get_by_assignment(assignment, preload \\ [])

get_by_promotion(promotion, preload \\ [])

get_by_submission(submission, preload \\ [])

get_changeset(attrs \\ %{})

get_user_pools_by_currency(user, currency)

handle_exclusion(assignment, items)

import_participant_reward(participant_id, amount, session_key, currency)

import_participant_reward_exists?(participant_id, session_key)

list(opts \\ [])

list_by_assignments(assignment_ids, preload)

list_by_budget(model, preload \\ [])

list_by_participant(user, opts \\ [])

Returns the list of adverts where the user has the specified role

list_by_pools_and_submission_status(pools, submission_status, opts \\ [])

list_by_status(status, opts \\ [])

list_excluded_adverts(adverts, preload \\ [])

list_excluded_user_ids(advert_ids)

list_owners(advert, preload \\ [])

list_submitted(pool, opts \\ [])

Returns the list of adverts submitted at least once.

list_tools(advert)

list_tools(advert, schema)

mark_expired_debug(force \\ false)

Marks expired tasks in online adverts based on updated_at and estimated duration. If force is true (for debug purposes only), all pending tasks will be marked as expired.

mark_expired_debug(map, force)

Marks expired tasks in given advert

open_spot_count(arg1)

payout_participant(assignment, user)

query_user_pools_by_currency(user, currency)

ready?(id)

reward_amount(model)

reward_value(assignment)

rewarded_amount(assignment, user)

submission_updated(arg1)

update(changeset)

Updates a get_advert.

Examples

iex> update(get_advert, %{field: new_value})
{:ok, %Advert.Model{}}

iex> update(get_advert, %{field: bad_value})
{:error, %Ecto.Changeset{}}

update(advert, attrs)

user_has_currency?(user, currency)

validate_open(advert)

validate_open(assignment, user)